Add tests for boolean values; Clean up JIT function module

+++ dust-lang/src/compiler/mod.rs
@@ -1655,7 +1655,7 @@ impl<'a> ChunkCompiler<'a> {
             }
         }
 
-        let mut destination_register = (last_item_register + 1) as u16;
+        let mut destination_register = last_item_register + 1;
         let reordered_instructions_count = instructions_to_reorder.len() as u16;
 
         for mut instruction in instructions_to_reorder {
